% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_alpha.R
\name{calc_alpha}
\alias{calc_alpha}
\title{Calculate the probability a type G customer is delayed while a restricted server is free}
\usage{
calc_alpha(m, n, prob_vec, K)
}
\arguments{
\item{m}{:= number of restricted servers}

\item{n}{:= the number of general use servers}

\item{prob_vec}{:= state probability matrix}
}
\value{
probability there is a delay of a general customer when a restricted server is free
}
\description{
Calculate the probability a type G customer is delayed while a restricted server is free
}
