% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_B_ki.R
\name{Calc_B_ki}
\alias{Calc_B_ki}
\title{Calculating B matrices}
\usage{
Calc_B_ki(K, m, n, lam, lam_R, lam_G, mu_R, mu_G, p)
}
\arguments{
\item{K}{:= truncation parameter}

\item{m}{:= number of restricted servers}

\item{n}{:= number of general servers}

\item{lam}{:= arrival rate of all customers}

\item{lam_R}{:= arrival rate of customers of type restricted}

\item{lam_G}{:= arrival rate of customers of type general}

\item{mu_R}{:= restricted server rate}

\item{mu_G}{:= general server rate}

\item{p}{:= percentage of customers that are of type restricted}
}
\value{
matrix B of the states with boundary behavior
}
\description{
Calculating B matrices
}
\examples{
#' K <- 13
m <- 10
n <- 5
lambda <- .2
p <-.9
lambda_aux <- (p-1) * lambda
lambda_p <- p * lambda
mu_g <-.4
mu_r <-.5
}
