% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_L_q_g.R, R/calc_L_q_r.R
\name{calc_L_q_r}
\alias{calc_L_q_r}
\title{Calculate the steady-state number of customers in the "general queue"}
\usage{
calc_L_q_r(K, m, prob_vec)

calc_L_q_r(K, m, prob_vec)
}
\arguments{
\item{K}{:=truncation parameter}

\item{m}{:= number of restricted servers}

\item{prob_vec}{:= state probability vector}

\item{n}{:= number of general servers}
}
\value{
L_q_g value

L_q_r value
}
\description{
Calculate the steady-state number of customers in the "general queue"

Calculate the steady-state number of customers in the "restricted queue"
}
\examples{
K <- 14
n <- 2
prob_vec <- []
calc_L_q_r(K,m,prob_vec)
K <- 14
m <- 2
prob_vec <- []
calc_L_q_r(K,m,prob_vec)
}
