% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_B_g.R
\name{calc_p_B_g}
\alias{calc_p_B_g}
\title{Calculate The probability that a type G customer has a positive delay}
\usage{
calc_p_B_g(K, n, prob_vec)
}
\arguments{
\item{K}{:= truncation parameter}

\item{n}{:= the number of general use servers}

\item{prob_vec}{:= state probability matrix}
}
\value{
p_B_g value := the probability all G servers are busy
}
\description{
Calculate The probability that a type G customer has a positive delay
}
