% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_B_r.R
\name{calc_p_B_r}
\alias{calc_p_B_r}
\title{Calculate the probability that a type R customer has a positive delay}
\usage{
calc_p_B_r(m, n, K, prob_vec)
}
\arguments{
\item{m}{:= number of restricted servers}

\item{n}{:= the number of general use servers}

\item{K}{:= truncation parameter}

\item{prob_vec}{:= state probability matrix}
}
\value{
p_B_r value
}
\description{
Calculate the probability that a type R customer has a positive delay
}
