% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_R.R
\name{calc_R}
\alias{calc_R}
\title{Solve for R}
\usage{
calc_R(A, K, n, tolerance = 1e-04)
}
\arguments{
\item{A}{:= A_m matrix that represents non-boundary behavior}

\item{K}{:= truncation parameter}

\item{n}{:= number of general servers}

\item{tolerance}{:= max difference between R iterations}
}
\value{
R matrix
}
\description{
Solve for R
}
\examples{
K <- 13
s <- 5
r <- 10
lambda <- .2
p <-.9
lambda_aux <- (p-1) * lambda
lambda_p <- p * lambda
mu_p <-.4
mu_aux <-.5
A <- Calc_Am (K,s,r,lambda,lambda_aux,lambda_p,mu_p,mu_aux, p)
Calc_R(A, K,s)
}
