% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rho_g.R
\name{calc_rho_g}
\alias{calc_rho_g}
\title{Calculate The probability tbat a type R customer has a positive delay}
\usage{
calc_rho_g(K, p, X, n, m, p_B_R, p_B_G, mu_R, mu_G, lam_r, lam_g)
}
\arguments{
\item{n}{:= the number of general use servers}

\item{m}{:= number of restricted servers}

\item{p_B_R}{:= probability of delay for R customers}

\item{p_B_G}{:= probability of delay for G customers}

\item{mu_R}{:= Restricted server service rate}

\item{mu_G}{:= General server service rate}

\item{lam_r}{:= arrival rate of R customers}

\item{lam_g}{:= arrival rate of G customers}
}
\value{
utilization of general servers
}
\description{
Calculate The probability tbat a type R customer has a positive delay
}
