% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rho_r.R
\name{calc_rho_r}
\alias{calc_rho_r}
\title{Calculate The probability tbat a type R customer has a positive delay}
\usage{
calc_rho_r(K, p, n, m, p_B_R, p_B_G, mu_R, mu_G, lam_r, prob_vec)
}
\arguments{
\item{n}{:= the number of general use servers}

\item{m}{:= number of restricted servers}

\item{p_B_R}{:= probability of delay for R customers}

\item{p_B_G}{:= probability of delay for G customers}

\item{mu_R}{:= Restricted server service rate}

\item{mu_G}{:= General server service rate}

\item{lam_r}{:= arrival rate of R customers}
}
\value{
Utilization of restricted servers
}
\description{
Calculate The probability tbat a type R customer has a positive delay
}
