% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_X.R
\name{calc_X}
\alias{calc_X}
\title{Solve for stationary probability vector X}
\usage{
calc_X(K, m, n, A, B, R, tolerance = 1e-05, x_i_thres = 1e-05)
}
\arguments{
\item{K}{:= truncation parameter}

\item{m}{:= m matrix}

\item{n}{:= number of general servers}

\item{A}{:=  A_m matrix that represents non-boundary behavior}

\item{B}{:= matrix B of the states with boundary behavior}
}
\value{
stationary probability vector X
}
\description{
Solve for stationary probability vector X
}
\examples{
K <- 13
m <- 10
n <- 5
lambda <- .2
p <-.9
lambda_aux <- (p-1) * lambda
lambda_p <- p * lambda
mu_p <-.4
mu_aux <-.5
A <- Calc_Am(K,s,r,lambda,lambda_aux,lambda_p,mu_p,mu_aux, p)
m <- Calc_R(A, K,s)
B <-  Calc_Bmn(K,s,r,lambda,lambda_aux,lambda_p,mu_p,mu_aux, p)
X_i <- Calc_X(K,s,r, A,B,R)
}
