% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_system.R
\name{check_system}
\alias{check_system}
\title{Check for steady state system conditions}
\usage{
check_system(lam, n, m, p, mu_g, mu_r)
}
\arguments{
\item{lam}{:= arrival rate of all customers}

\item{n}{:= the number of general use servers}

\item{m}{:= the number of restricted servers}

\item{p}{:= the probability a customer requires restricted service}

\item{mu_g}{:= the service rate of general servers}

\item{mu_r}{:= the service rate of restricted servers}
}
\value{
string indicating steady state exists or not
}
\description{
Check for steady state system conditions
}
