% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensecqueue.R
\name{gensecqueue}
\alias{gensecqueue}
\title{Calculate performance metrics of a queueing system with General and Restricted Servers}
\usage{
gensecqueue(lam, n, m, p, mu_g, mu_r, N = 1 + n, tolerance = 1e-06)
}
\arguments{
\item{lam}{:= arrival rate of all customers}

\item{n}{:= the number of general use servers}

\item{m}{:= the number of restricted servers}

\item{p}{:= the probability a customer requires restricted service}

\item{mu_g}{:= the service rate of general servers}

\item{mu_r}{:= the service rate of restricted servers}
}
\value{
data frame of key performance metrics
}
\description{
Calculate performance metrics of a queueing system with General and Restricted Servers
}
