% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_row_G.R
\name{last_row_G}
\alias{last_row_G}
\title{Append last row to truncated G matrix}
\usage{
last_row_G(G, K, R, m)
}
\arguments{
\item{G}{:= truncated Generator matrix}

\item{K}{:= max queue length}

\item{R}{:= intermediate matrix}

\item{m}{:= number of servers}
}
\description{
Append last row to truncated G matrix
}
\examples{
G <- matrix
K <- 10
R <- matrix
m <- 3
G_final <- last_row_G(G, K, R, m)
}
