% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_X.R
\name{reformat_X}
\alias{reformat_X}
\title{Reformat probability matrix from X\link{k} to X\link{i,j}}
\usage{
reformat_X(X, matrix_size)
}
\arguments{
\item{X}{:= probability matrix}

\item{matrix_size}{:= size of reformatted matrix}
}
\value{
:= re-sized probability matrix
}
\description{
Reformat probability matrix from X\link{k} to X\link{i,j}
}
