% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_K.R
\name{Solve_K}
\alias{Solve_K}
\title{Solve for Truncation parameter K ######}
\usage{
Solve_K(m, n, lambda, mu_g, mu_r, p, N)
}
\arguments{
\item{m}{:= the number of restricted servers}

\item{n}{:= the number of general use servers}

\item{lambda}{:= arrival rate of all customers}

\item{mu_g}{:= the service rate of general servers}

\item{mu_r}{:= the service rate of restricted server}

\item{p}{:= the probability a customer requires restricted service}

\item{N}{:= temporary starting truncation parameter}
}
\value{
final K truncation parameter with key metric stability
}
\description{
Solve for Truncation parameter K ######
}
\examples{
r <- 10
s <- 5
lambda <- .2
p <-.9
lambda_aux <- (p-1) * lambda
lambda_p <- p * lambda
mu_p <-.4
mu_aux <-.5
K <- Solve_K(s,r,lambda,lambda_aux,lambda_p,mu_p,mu_aux, p, N = s+1)
}
