% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_prob_matrix.R
\name{solve_prob_matrix}
\alias{solve_prob_matrix}
\title{Solve Probability Matrix}
\usage{
solve_prob_matrix(G, b)
}
\arguments{
\item{G}{:= Truncated Generator Matrix}

\item{b}{:= Right hand side values}
}
\value{
probability matrix
}
\description{
Solve Probability Matrix
}
\examples{
add example from test
G <- matrix(0)
b <- matrix(0)
X <- solve_prob_matrix(G,b)
}
