% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_G.R
\name{trunc_G}
\alias{trunc_G}
\title{Full Generator matrix}
\usage{
trunc_G(K, n, m, A, B, R)
}
\arguments{
\item{K}{:= truncation parameter}

\item{n}{:= number of general servers}

\item{m}{:= number of restricted servers}

\item{A}{:=  A_m matrix that represents non-boundary behavior}

\item{B}{:= matrix B of the states with boundary behavior}

\item{R}{:= R matrix}
}
\value{
Full Generator matrix
}
\description{
Full Generator matrix
}
\examples{
K <- 13
r <- 10
s <- 5
lambda <- .2
p <-.9
lambda_aux <- (p-1) * lambda
lambda_p <- p * lambda
mu_p <-.4
mu_aux <-.5
A <- Calc_A_k(K,s,r,lambda,lambda_aux,lambda_p,mu_p,mu_aux, p)
R <- Calc_R(A, K,s)
B <-  Calc_B_ki(K,s,r,lambda,lambda_aux,lambda_p,mu_p,mu_aux, p)
trunc_G(K,s,r, A,B,R)
}
